football_stats/requirement.txt
psycopg2-binary
python-dotenv
sqlalchemy
google-cloud-sql-connector[pg8000]